tag 过滤 bugfix

+++ rocketmq-store/src/main/java/com/alibaba/rocketmq/store/DefaultMessageFilter.java
@@ -43,7 +43,7 @@ public class DefaultMessageFilter implements MessageFilter {
             return true;
         }
 
-        return subscriptionData.getCodeSet().contains(tagsCode);
+        return subscriptionData.getCodeSet().contains(tagsCode.intValue());
     }
 
 }
